refactor: move pull related functions to pull module
